+++ micrometer-jakarta10/src/main/java/io/micrometer/jakarta10/instrument/jms/JmsPublishObservationConvention.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.micrometer.core.instrument.binder.jms;
+package io.micrometer.jakarta10.instrument.jms;
 
 import io.micrometer.observation.Observation;
 import io.micrometer.observation.ObservationConvention;
