+++ micrometer-jakarta10/src/main/java/io/micrometer/jakarta10/instrument/jms/MessageProducerInvocationHandler.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micrometer.core.instrument.binder.jms;
+package io.micrometer.jakarta10.instrument.jms;
 
 import io.micrometer.common.lang.Nullable;
 import io.micrometer.observation.Observation;
