+++ micrometer-test/build.gradle
@@ -36,6 +36,7 @@ dependencies {
     implementation 'org.awaitility:awaitility'
 
     testImplementation project(':micrometer-observation-test')
+    testImplementation project(':micrometer-jakarta10')
 
     testImplementation 'org.jsr107.ri:cache-ri-impl'
